feat: add support for typed casts (#98)
* feat: add support for typed casts

* add tests; update type conversions; update PLK

+++ src/test/kotlin/org/partiql/scribe/utils/LocalSchema.kt
@@ -38,7 +38,7 @@ private object LocalSchema {
             "int8" -> PType.tinyint()
             "int16" -> PType.smallint()
             "int32" -> PType.integer()
-            "int64" -> PType.integer()
+            "int64" -> PType.bigint()
             "decimal" -> PType.decimal()
             "float32" -> PType.real()
             "float64" -> PType.doublePrecision()
